+++ canaille/app/flask.py
@@ -4,13 +4,13 @@ from urllib.parse import urlsplit
 from urllib.parse import urlunsplit
 
 from canaille.app import models
+from canaille.app.i18n import gettext as _
 from canaille.app.themes import render_template
 from flask import abort
 from flask import current_app
 from flask import g
 from flask import request
 from flask import session
-from flask_babel import gettext as _
 from werkzeug.routing import BaseConverter