+++ canaille/oidc/clients.py
@@ -4,6 +4,7 @@ from canaille.app import models
 from canaille.app.flask import permissions_needed
 from canaille.app.flask import render_htmx_template
 from canaille.app.forms import TableForm
+from canaille.app.i18n import gettext as _
 from canaille.app.themes import render_template
 from canaille.oidc.forms import ClientAddForm
 from flask import abort
@@ -12,7 +13,6 @@ from flask import flash
 from flask import redirect
 from flask import request
 from flask import url_for
-from flask_babel import gettext as _
 from werkzeug.security import gen_salt